feat: Add --init option to update command
