+++ internal/cmd/inittemplatefuncs.go
@@ -1,0 +1,54 @@
+package cmd
+
+import (
+	"fmt"
+	"os"
+	"strconv"
+	"strings"
+
+	"golang.org/x/term"
+)
+
+func (c *Config) promptBool(field string) bool {
+	value, err := parseBool(c.promptString(field))
+	if err != nil {
+		returnTemplateError(err)
+		return false
+	}
+	return value
+}
+
+func (c *Config) promptInt(field string) int64 {
+	value, err := strconv.ParseInt(c.promptString(field), 10, 64)
+	if err != nil {
+		returnTemplateError(err)
+		return 0
+	}
+	return value
+}
+
+func (c *Config) promptString(field string) string {
+	value, err := c.readLine(fmt.Sprintf("%s? ", field))
+	if err != nil {
+		returnTemplateError(err)
+		return ""
+	}
+	return strings.TrimSpace(value)
+}
+
+func (c *Config) stdinIsATTY() bool {
+	file, ok := c.stdin.(*os.File)
+	if !ok {
+		return false
+	}
+	return term.IsTerminal(int(file.Fd()))
+}
+
+func (c *Config) writeToStdout(args ...string) string {
+	for _, arg := range args {
+		if _, err := c.stdout.Write([]byte(arg)); err != nil {
+			panic(err)
+		}
+	}
+	return ""
+}
